Development branch

+++ Manager.Infrastructure/Repositories/Models/OrdersRepo.cs
@@ -1,5 +1,4 @@
 ﻿using Dapper;
-using Manager.Application.Interfaces;
 using Manager.Application.Repository.Interfaces;
 using Manager.Domain.Models;
 using Manager.Infrastructure.Connection;
@@ -7,7 +6,7 @@ using Npgsql;
 
 namespace Manager.Infrastructure.Repositories.Models
 {
-    public class OrdersRepo :IOrdersRepository
+    public class OrdersRepo : IOrdersRepository
     {
         private readonly string _connection;
         public OrdersRepo()
@@ -17,17 +16,41 @@ namespace Manager.Infrastructure.Repositories.Models
 
         public async Task<bool> DeleteByIdAsync(int orderId)
         {
+            if (await new OrderProductsRepo().DeleteByOrderIdAsync(orderId))
+            {
+                using (NpgsqlConnection conn = new NpgsqlConnection(_connection))
+                {
+                    await conn.OpenAsync();
 
+                    string cmdText = @"delete from orders where order_id = @id";
+                    if (await conn.ExecuteAsync(cmdText, new { id = orderId }) > 0) return true;
+                    else return false;
+                }
+            }
+            else return false;
+            
+        }
+        public async Task<int> InsertOrderReturnId(Order order)
+        {
             using (NpgsqlConnection conn = new NpgsqlConnection(_connection))
             {
                 await conn.OpenAsync();
+                string cmdText = @"insert into orders(
+                           order_table, order_date, is_completed)
+                           values (@OrderTable, @OrderDate,
+                           @IsCompleted) returning order_id;";
 
-                string cmdText = @"delete from orders where order_id = @id";
-                if (await conn.ExecuteAsync(cmdText, new { id = orderId }) > 0) return true;
-                else return false;
-
+                return await conn.ExecuteScalarAsync<int>(cmdText,
+                    new
+                    {
+                        OrderTable = order.OrderTable,
+                        OrderDate = DateTime.Now,
+                        IsCompleted = order.IsCompleted
+                    });
             }
         }
+
+
 
         public async Task<List<Order>> GetAllAsync()
         {
@@ -42,13 +65,12 @@ namespace Manager.Infrastructure.Repositories.Models
                     orders.Add(new Order
                     {
                         OrderId = reader.GetInt32(0),
-                        Waiter = await new WaiterRepo().GetByIdAsync(reader.GetInt32(1)),
-                        OrderTable = reader.GetInt32(2),
-                        OrderDate = reader.GetDateTime(3),
-                        IsCompleted = reader.GetBoolean(4)
+                        OrderTable = reader.GetInt32(1),
+                        OrderDate = reader.GetDateTime(2),
+                        IsCompleted = reader.GetBoolean(3)
                     });
                 }
-              
+
                 return orders;
             }
         }
@@ -57,20 +79,25 @@ namespace Manager.Infrastructure.Repositories.Models
         {
             using (NpgsqlConnection conn = new NpgsqlConnection(_connection))
             {
-                Order orders = new Order();
+
                 await conn.OpenAsync();
                 string cmdText = @"select * from orders where order_id = @id;";
-                var reader = await conn.ExecuteReaderAsync(cmdText, new { id = orderId });
+                var reader = await conn.ExecuteReaderAsync(cmdText, new{id = orderId});
+                Order orders = new Order();
+                while(await reader.ReadAsync())
+                {
+                    orders.OrderId = reader.GetInt32(0);
+                    orders.OrderTable = reader.GetInt32(1);
+                    orders.OrderDate = reader.GetDateTime(2);
+                    orders.IsCompleted = reader.GetBoolean(3);
 
-                orders.OrderId = reader.GetInt32(0);
-                orders.Waiter = await new WaiterRepo().GetByIdAsync(reader.GetInt32(1));
-                orders.OrderTable = reader.GetInt32(2);
-                orders.OrderDate = reader.GetDateTime(3);
-                orders.IsCompleted = reader.GetBoolean(4);
+                }
 
                 return orders;
             }
+
         }
+
 
         public async Task<bool> InsertAsync(Order orders)
         {
@@ -78,15 +105,15 @@ namespace Manager.Infrastructure.Repositories.Models
             {
                 await conn.OpenAsync();
                 string cmdText = @"insert into orders(
-                                   waiter_id, order_table, order_date, is_completed)
-                                   values (@Waiter, @OrderTable, @OrderDate,
+                                   order_table, order_date, is_completed)
+                                   values (@OrderTable, @OrderDate,
                                    @IsCompleted);";
-                if (await conn.ExecuteAsync(cmdText, 
-                        new {
-                                Waiter = orders.Waiter.WaiterId,
-                                OrderTable = orders.OrderTable,
-                                OrderDate = orders.OrderDate,
-                                IsCompleted = orders.IsCompleted
+                if (await conn.ExecuteAsync(cmdText,
+                        new
+                        {
+                            OrderTable = orders.OrderTable,
+                            OrderDate = DateTime.Now,
+                            IsCompleted = orders.IsCompleted
                         }) > 0) return true;
                 else return false;
             }
@@ -98,21 +125,21 @@ namespace Manager.Infrastructure.Repositories.Models
             {
                 await conn.OpenAsync();
                 string cmdText = @"update orders set 
-                                   waiter_id = @Waiter,
                                    order_table = @OrderTable,
                                    order_date = @OrderDate,
                                    is_completed = @IsCompleted
-                                   where order_id = @OrderId,";
-                if (await conn.ExecuteAsync(cmdText, 
-                        new {
-                                OrderId = orders.OrderId,
-                                Waiter = orders.Waiter.WaiterId,
-                                OrderTable = orders.OrderTable,
-                                OrderDate = orders.OrderDate,
-                                IsCompleted = orders.IsCompleted,
-                            }) > 0) return true;
+                                   where order_id = @OrderId;";
+                if (await conn.ExecuteAsync(cmdText,
+                        new
+                        {
+                            OrderId = orders.OrderId,
+                            OrderTable = orders.OrderTable,
+                            OrderDate = orders.OrderDate,
+                            IsCompleted = orders.IsCompleted,
+                        }) > 0) return true;
                 else return false;
             }
         }
     }
 }
+
